requirements.txt
# Core ML/AI Dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.1
pinecone>=5.0.0
openai>=1.0.0

# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
arxiv>=1.4.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0

# Multi-Agent Framework
crewai>=0.20.0
autogen-agentchat>=0.2.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
python-multipart>=0.0.6
pydantic>=2.0.0 